% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biplot.PCovR.R
\name{Biplot.PCovR}
\alias{Biplot.PCovR}
\title{This function generates a biplot for the results of the PCovR analysis.}
\usage{
Biplot.PCovR(X, Y, PcovR)
}
\arguments{
\item{X}{A matrix containing the data for the X variables.}

\item{Y}{A matrix containing the data for the Y variables.}

\item{PcovR}{A list containing the results of the PCovR analysis.}
}
\value{
A list containing the biplot information.
}
\description{
This function generates a biplot for the results of the PCovR analysis.
}
