% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biplot.PCovR.Weights.R
\name{Biplot.PCovR.Weights}
\alias{Biplot.PCovR.Weights}
\title{PCovR Regression Biplot}
\usage{
Biplot.PCovR.Weights(PcovR)
}
\arguments{
\item{PcovR}{A list containing the results of the PCovR analysis.}
}
\value{
A list containing the biplot information.
}
\description{
This function generates a regression biplot to examine the relationship
between the response and explanatory variables.
}
